requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
pyodbc==5.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
flasgger==0.9.7.1
bcrypt==4.1.2
Werkzeug==3.0.1

